Previous results (seed = 42, regions = 25, mapSize = 65, domain = {1,2,3,4})
```
__________________________
Ac3
First solution:
        First solution found in: 19 ms
        Nodes visited to find first solution: 26
All solutions:
        Nodes visited: 125876
        Time duration: 2405 ms
        Found: 24024
__________________________
Backtracking
First solution:
        First solution found in: 3 ms
        Nodes visited to find first solution: 200
All solutions:
        Nodes visited: 884052
        Time duration: 718 ms
        Found: 24024
__________________________
ForwardChecking
First solution:
        First solution found in: 5 ms
        Nodes visited to find first solution: 34
All solutions:
        Nodes visited: 363925
        Time duration: 1101 ms
        Found: 24024

__________________________________________

__________________________
Ac3
First solution:
        First solution found in: 15 ms
        Nodes visited to find first solution: 26
All solutions:
        Nodes visited: 28
        Time duration: 2 ms
        Found: 1
__________________________
Backtracking
First solution:
        First solution found in: 2 ms
        Nodes visited to find first solution: 528
All solutions:
        Nodes visited: 1337
        Time duration: 1 ms
        Found: 1
__________________________
ForwardChecking
First solution:
        First solution found in: 4 ms
        Nodes visited to find first solution: 41
All solutions:
        Nodes visited: 169
        Time duration: 1 ms
        Found: 1
```


New results (seed = 42, regions = 25, mapSize = 65, domain = {1,2,3,4})
```__________________________
Ac3
First solution:
        First solution found in: 15 ms
        Nodes visited to find first solution: 25
All solutions:
        Nodes visited: 108232
        Time duration: 1935 ms
        Found: 24024
__________________________
Backtracking
First solution:
        First solution found in: 3 ms
        Nodes visited to find first solution: 200
All solutions:
        Nodes visited: 884052
        Time duration: 540 ms
        Found: 24024
__________________________
ForwardChecking
First solution:
        First solution found in: 4 ms
        Nodes visited to find first solution: 25
All solutions:
        Nodes visited: 110344
        Time duration: 435 ms
        Found: 24024

__________________________________________

__________________________
Ac3
First solution:
        First solution found in: 13 ms
        Nodes visited to find first solution: 25
All solutions:
        Nodes visited: 26
        Time duration: 1 ms
        Found: 1
__________________________
Backtracking
First solution:
        First solution found in: 1 ms
        Nodes visited to find first solution: 83
All solutions:
        Nodes visited: 1507
        Time duration: 0 ms
        Found: 1
__________________________
ForwardChecking
First solution:
        First solution found in: 4 ms
        Nodes visited to find first solution: 26
All solutions:
        Nodes visited: 112
        Time duration: 0 ms
        Found: 1```